pyproject.toml
[tool.poetry]
name = "hourly-register"
version = "0.1.0"
description = "Fill up monthly hourly report in compliance with Spanish Law"
authors = ["Alberto Camara <camara@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
jinja2 = "^2.11.2"
pyyaml = "^5.3.1"
rope = "^0.17.0"

[tool.poetry.dev-dependencies]
pytest = "^6.0.2"
pytest-cov = "^2.10.1"
pytest-mock = "^3.3.1"
pylint = "^2.6.0"
black = "^20.8b1"
pdbpp = "^0.10.2"
ipython = "^7.18.1"
mypy = "^0.782"
bandit = "^1.6.2"
rope = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
